requirements.txt
flask
web3
openai
python-dotenv